# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Shepherd API
  description: |
    API REST pour le système de monitoring IoT Smart Shepherd.
    Gère l'authentification, les animaux, la télémétrie, les géofences, les notifications et les utilisateurs.
  version: 1.0.0
  contact:
    name: Smart Shepherd Team
    email: support@smartshepherd.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Serveur de développement
  - url: https://api.smartshepherd.com/v1
    description: Serveur de production
...
paths:
...
  /sheep:
    get:
      tags:
        - Animaux
      summary: Lister tous les animaux
      description: Retourne la liste complète du troupeau avec pagination
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/HealthStatus'
      responses:
        '200':
          description: Liste des animaux récupérée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Sheep'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
              example:
                success: true
                data:
                  - id: "64f8a1b2c3d4e5f6a7b8c9d0"
                    sheepId: "SHEEP001"
                    breed: "Merinos"
                    age: 3
                    weight: 75.5
                    gender: "male"
                    healthStatus: "Good"
                    location:
                      lat: 33.885
                      lng: -5.54
                    deviceId: "C001"
                    medicalHistory: []
                    createdAt: "2023-09-15T10:30:00.000Z"
                pagination:
                  page: 1
                  limit: 20
                  total: 150
                  pages: 8
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Animaux
      summary: Ajouter un nouvel animal
      description: Crée un nouvel enregistrement d'animal dans le troupeau
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SheepCreate'
            example:
              sheepId: "SHEEP002"
              breed: "Suffolk"
              age: 2
              weight: 68.2
              gender: "female"
              deviceId: "C002"
      responses:
        '201':
          description: Animal créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Animal créé avec succès"
                  data:
                    $ref: '#/components/schemas/Sheep'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /sheep/{id}:
    get:
      tags:
        - Animaux
      summary: Obtenir un animal spécifique
      description: Retourne les détails d'un animal par son ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SheepId'
      responses:
        '200':
          description: Animal récupéré avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Sheep'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Animaux
      summary: Mettre à jour un animal
      description: Modifie les informations d'un animal existant
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SheepId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SheepUpdate'
            example:
              breed: "Merinos"
              age: 4
              weight: 78.0
              healthStatus: "Warning"
      responses:
        '200':
          description: Animal mis à jour avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Animal mis à jour avec succès"
                  data:
                    $ref: '#/components/schemas/Sheep'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Animaux
      summary: Supprimer un animal
      description: Supprime un animal du système (soft delete)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SheepId'
      responses:
        '200':
          description: Animal supprimé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Animal supprimé avec succès"
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /sheep/{id}/location:
    put:
      tags:
        - Animaux
      summary: Mettre à jour la position d'un animal
      description: Enregistre la nouvelle position GPS d'un animal
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SheepId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationUpdate'
            example:
              lat: 33.886
              lng: -5.541
              accuracy: 5.2
              timestamp: "2023-09-15T10:30:00.000Z"
      responses:
        '200':
          description: Position mise à jour avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Position mise à jour avec succès"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /sheep/{id}/medical:
    post:
      tags:
        - Animaux
      summary: Ajouter un enregistrement médical
      description: Ajoute une nouvelle entrée dans l'historique médical d'un animal
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SheepId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicalRecord'
            example:
              type: "vaccination"
              description: "Vaccin contre la fièvre aphteuse"
              veterinarian: "Dr. Martin"
              date: "2023-09-15T10:30:00.000Z"
              notes: "Réaction normale observée"
      responses:
        '201':
          description: Enregistrement médical ajouté avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Enregistrement médical ajouté avec succès"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # TÉLÉMÉTRIE ENDPOINTS
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT d'authentification
  parameters:
    Page:
      in: query
      name: page
      description: Numéro de page pour la pagination
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      in: query
      name: limit
      description: Nombre d'éléments par page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Search:
      in: query
      name: search
      description: Terme de recherche
      required: false
      schema:
        type: string
    HealthStatus:
      in: query
      name: healthStatus
      description: Filtre par statut de santé
      required: false
      schema:
        type: string
        enum: [Good, Warning, Critical]
    SheepId:
      in: path
      name: id
      description: ID de l'animal
      required: true
      schema:
        type: string
        example: "64f8a1b2c3d4e5f6a7b8c9d0"
...
  schemas:
...
    Sheep:
      type: object
      properties:
        id:
          type: string
          example: "64f8a1b2c3d4e5f6a7b8c9d0"
        sheepId:
          type: string
          description: Identifiant unique de l'animal
          example: "SHEEP001"
        breed:
          type: string
          description: Race de l'animal
          example: "Merinos"
        age:
          type: integer
          description: Âge en années
          example: 3
        weight:
          type: number
          format: float
          description: Poids en kg
          example: 75.5
        gender:
          type: string
          enum: [male, female]
          example: "male"
        healthStatus:
          type: string
          enum: [Good, Warning, Critical]
          example: "Good"
        location:
          $ref: '#/components/schemas/Location'
        deviceId:
          type: string
          description: ID du collier IoT
          example: "C001"
        medicalHistory:
          type: array
          items:
            $ref: '#/components/schemas/MedicalRecord'
        createdAt:
          type: string
          format: date-time
          example: "2023-09-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-09-20T14:25:00.000Z"
        isActive:
          type: boolean
          example: true
    SheepCreate:
      type: object
      required:
        - sheepId
        - breed
        - age
        - weight
        - gender
        - deviceId
      properties:
        sheepId:
          type: string
          example: "SHEEP002"
        breed:
          type: string
          example: "Suffolk"
        age:
          type: integer
          example: 2
        weight:
          type: number
          format: float
          example: 68.2
        gender:
          type: string
          enum: [male, female]
          example: "female"
        deviceId:
          type: string
          example: "C002"
    SheepUpdate:
      type: object
      properties:
        breed:
          type: string
          example: "Merinos"
        age:
          type: integer
          example: 4
        weight:
          type: number
          format: float
          example: 78.0
        healthStatus:
          type: string
          enum: [Good, Warning, Critical]
          example: "Warning"
    Location:
      type: object
      properties:
        lat:
          type: number
          format: float
          description: Latitude
          example: 33.885
        lng:
          type: number
          format: float
          description: Longitude
          example: -5.54
        accuracy:
          type: number
          format: float
          description: Précision GPS en mètres
          example: 3.5
        timestamp:
          type: string
          format: date-time
          example: "2023-09-15T10:30:00.000Z"
    LocationUpdate:
      type: object
      required:
        - lat
        - lng
      properties:
        lat:
          type: number
          format: float
          example: 33.886
        lng:
          type: number
          format: float
          example: -5.541
        accuracy:
          type: number
          format: float
          example: 5.2
        timestamp:
          type: string
          format: date-time
          example: "2023-09-15T10:30:00.000Z"
    MedicalRecord:
      type: object
      required:
        - type
        - description
        - veterinarian
        - date
      properties:
        type:
          type: string
          enum: [vaccination, treatment, checkup, surgery, other]
          example: "vaccination"
        description:
          type: string
          example: "Vaccin contre la fièvre aphteuse"
        veterinarian:
          type: string
          example: "Dr. Martin"
        date:
          type: string
          format: date-time
          example: "2023-09-15T10:30:00.000Z"
        notes:
          type: string
          example: "Réaction normale observée"
        medications:
          type: array
          items:
            type: string
          example: ["Vaccin", "Anti-inflammatoire"]

    # Télémétrie Schemas
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        pages:
          type: integer
          example: 8
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Les données fournies sont invalides"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: "email"
                  message:
                    type: string
                    example: "L'email est invalide"
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Les données fournies sont invalides"
              details:
                - field: "email"
                  message: "L'email est invalide"
    Unauthorized:
      description: Non autorisé
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "Token JWT manquant ou invalide"
...
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "NOT_FOUND"
              message: "La ressource demandée n'existe pas"
    Conflict:
      description: Conflit de données
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "CONFLICT"
              message: "Un utilisateur avec cet email existe déjà"
...